UI changes and local DB constraints

+++ EnparadigmTask/Source/Project/ViewModel/WeatherViewModel.swift
@@ -17,27 +17,27 @@ class WeatherViewModel {
     public var completionHandler: ((Bool, Error?, String?, String?) -> (Void))?
     
     func getData(city: String) {
-       if Reachability.isConnectedToNetwork() {
-        APIManager.shared.getWeatherData(city: city) { (result) in
-            switch result {
-            case let  .failure(_, title, subTitle):
-                if let completion = self.completionHandler {
-                    completion(false,nil, title, subTitle)
+        if Reachability.isConnectedToNetwork() {
+            APIManager.shared.getWeatherData(city: city) { (result) in
+                switch result {
+                case let  .failure(_, title, subTitle):
+                    if let completion = self.completionHandler {
+                        completion(false,nil, title, subTitle)
+                    }
+                case let .success(weatherData):
+                    self.handleData(data: weatherData)
                 }
-            case let .success(weatherData):
-                self.handleData(data: weatherData)
             }
+        } else{
+            self.getValueFromLocal()
+            
         }
-    } else{
-        self.getValueFromLocal()
-    
-    }
     }
     private func handleData(data: WeatherModel) {
         if let completion = self.completionHandler {
             model = data
             DispatchQueue.main.async {
-               let data : WeatherLocalData = WeatherLocalData(humidity: self.model?.main?.humidity ?? 0, city: self.model?.name ?? "", temperature: self.model?.main?.temp ?? 0.0, visibility: self.model?.visibility ?? 0, wind: self.model?.wind?.speed ?? 0.0, pressure: self.model?.main?.pressure ?? 0, weather_Desc: self.model?.weather?.first?.main ?? "")
+                let data : WeatherLocalData = WeatherLocalData(humidity: self.model?.main?.humidity ?? 0, city: self.model?.name ?? "", temperature: self.model?.main?.temp ?? 0.0, visibility: self.model?.visibility ?? 0, wind: self.model?.wind?.speed ?? 0.0, pressure: self.model?.main?.pressure ?? 0, weather_Desc: self.model?.weather?.first?.main ?? "")
                 self.saveWeatherData(data: data)
                 completion(true,nil, nil, nil)
             }
@@ -69,9 +69,13 @@ class WeatherViewModel {
     }
     func saveWeatherData(data: WeatherLocalData){
         let context = LocalDataBaseService.context
-         let entity = NSEntityDescription.entity(forEntityName: DBMembers.entityName, in: context)
+        let entity = NSEntityDescription.entity(forEntityName: DBMembers.entityName, in: context)
         let currentWeather = NSManagedObject(entity: entity!, insertInto: context)
-        context.delete(currentWeather)
+        let fetchRequest = NSFetchRequest<NSFetchRequestResult>(entityName: DBMembers.entityName)
+        let results = try? context.fetch(fetchRequest)
+        if results?.count ?? 0 > 1{
+            context.delete(currentWeather)
+        }
         currentWeather.setValue(data.weather_Desc, forKey: DBMembers.weather_Desc)
         currentWeather.setValue(data.visibility, forKey: DBMembers.visibility)
         currentWeather.setValue(data.temperature, forKey: DBMembers.temperature)
@@ -89,29 +93,25 @@ class WeatherViewModel {
         
     }
     func getValueFromLocal(){
-//        DispatchQueue.main.async {
+        DispatchQueue.main.async {
             let context = LocalDataBaseService.context
             let fetchRequest = NSFetchRequest<NSFetchRequestResult>(entityName: DBMembers.entityName)
-            
-//            do {
+            do {
                 let results = try? context.fetch(fetchRequest)
                 let obtainedResults = results as! [NSManagedObject]
-//                let firstResult = obtainedResults.first
-//                let weather_Desc = firstResult?.value(forKey: DBMembers.weather_Desc)
-//                let visibility = firstResult?.value(forKey: DBMembers.visibility)
-//                let temperature = firstResult?.value(forKey: DBMembers.temperature)
-//                let pressure = firstResult?.value(forKey: DBMembers.pressure)
-//                let humidity = firstResult?.value(forKey: DBMembers.humidity)
-//                let city = firstResult?.value(forKey: DBMembers.city)
-//                let wind = firstResult?.value(forKey: DBMembers.wind)
-//                self.localModel = WeatherLocalData(humidity: humidity as? Int, city: city as? String, temperature: temperature as? Double, visibility: visibility as? Int, wind: wind as? Double, pressure: pressure as? Int, weather_Desc: weather_Desc as? String)
-//                print("myValue: \(String(describing: wind))")
-        print(obtainedResults)
-        print(obtainedResults)
-
-//            } catch {
-//
-//            }
-//        }
+                let firstResult = obtainedResults.first
+                let weather_Desc = firstResult?.value(forKey: DBMembers.weather_Desc)
+                let visibility = firstResult?.value(forKey: DBMembers.visibility)
+                let temperature = firstResult?.value(forKey: DBMembers.temperature)
+                let pressure = firstResult?.value(forKey: DBMembers.pressure)
+                let humidity = firstResult?.value(forKey: DBMembers.humidity)
+                let city = firstResult?.value(forKey: DBMembers.city)
+                let wind = firstResult?.value(forKey: DBMembers.wind)
+                self.localModel = WeatherLocalData(humidity: humidity as? Int, city: city as? String, temperature: temperature as? Double, visibility: visibility as? Int, wind: wind as? Double, pressure: pressure as? Int, weather_Desc: weather_Desc as? String)
+                
+            } catch(let error) {
+                print(error.localizedDescription)
+            }
+        }
     }
 }
